Refactor: Extract shared filter code and rename FilterBar to DesktopFilters
